bvt: use libXML parsing and XIncludes instead of hand-rolled implementations.
Instead of <include> tags that are parsed by hand, instead use
the XInclude standard to do it. This replaces ~200 lines of code
with a single libXML call.

Also, when doing the "treat bvt-xml as include" bit in op-it-list,
we can easily just use libXML to do the heavy lifting, which is
approximately 100x more solid than using sed.

Additionally, we update the helloworld XML files to test a bunch
of this simply by having includes.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>