Adding the validate_side_acivated function in pnor image flash procedure.
By using this patch, we will get BMC and PNOR Version side info(Golden or Primary) before and after
PNOR FW Flash. And also making sure doing pnor flash when system is in primary side

And also increased the timeout to 60 sec's for waiting to BMC prompt, As it is failing when running in jenkins.