rename LPAR to Host
The LPAR terminology is from PowerVM/hardware where the machine is
partitioned up into Logical Partitions.

However, this is not what we are testing here, instead, we're testing
the Host. It makes more sense to refer to the Host (as in, the host
processor, rather than the service processor/BMC).

So, this patch is a giant string replace.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>