Get machine to run on from machines.xml and rework test configuration
This patch reworks how we work out where to run a test.

Instead of the old way of having to specify all of the details of a
machine on the command line, and not having flexibility to specify
that different machines may have different capabilities (e.g. AMI BMC,
OpenBMC, FSP, simulator, number of processors etc) we move to
a machines.xml configuration file that lists out all machines in
your test setup, and then the ./run script runs the tests on a machine.

This allows each lab to have their own machines.xml

We also move configuration file generation from run-op-bvt-setup over
to ./run so that we fill it out based on the --machine option rather than
on command line options.

This moves us closer to be able to support running the test suite against
non-BMC machines.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>