Add test for booting and scanning SOL for Petitboot/login
On IBM FSP based systems, the Boot Status IPMI sensor doesn't exist,
so we have to resort to using expect to scan the Serial-Over-LAN
output for getting to petitboot/linux login.

This patch adds a test to do that, which is also run on BMC systems
as it should also pass there.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>