Fix for spawning ipmi sol console right after power on.
BMC Based systems will boot immediately after a power on.

Till now we have sleep of 60s after power on of the system,
Due to this we have delay in spawning ipmi sol console, which
results in missing host boot isteps. Almost we are spawning after skiboot boots.

[SOL Session operational.  Use ~? for help]
[1031107993,5] BT: Interface initialized, IO 0x00e4
[1492723106,5] NVRAM: Size is 576 KB
[1718771809,3] NVRAM: Layout appears sane
[1718948751,5] CAPI: Preloading ucode 200ea
[1719070527,5] FLASH: Queueing preload of 2/200ea
[1719130226,5] FLASH: Queueing preload of 0/0
[1719450159,7] FFS: Partition map size: 0x1000
[1721108239,7] FLASH: CAPP partition has ECC
[1722936302,7] FLASH: flash subpartition eyecatcher CAPP
[1722938559,7] FLASH: flash found subpartition: 1 size: 36432 offset 4096
[1876337311,5] FLASH: Queueing preload of 1/0

This patch reduces the delays and helps spawning the ipmi sol console
at right time.

[SOL Session operational.  Use ~? for help]
 15.88086|HWAS|PRESENT> DIMM[03]=AAAAAAAAAAAAAAAA
 15.88087|HWAS|PRESENT> Membuf[04]=CCCC000000000000
 15.88087|HWAS|PRESENT> Proc[05]=C000000000000000
 15.96799|ISTEP  6. 6
 16.18807|ISTEP  6. 7
 20.00460|ISTEP  6. 8
 20.11694|ISTEP  6. 9
 23.03573|ISTEP  6.10
 23.06527|ISTEP  6.11
 24.85307|ISTEP  6.12

And also this patch elemenates the problem of host_sol.log getting over written
in a multi boot test(If test case boots the system for multiple times)

With this patch logs like this host_sol_20160816_2019.log will be saved for
all the boots with the right time stamp.

Signed-off-by: Pridhiviraj Paidipeddi <ppaidipe@linux.vnet.ibm.com>