Fix for running HMI tests on ubuntu.
Initial HMI testcases implemetation was tested on rhel 7.2 and PowerKVM, Now tested it
on ubuntu and fixed the necessary changes.

And also added change to gather opal message logs for further debug during
failure of any HMI test.

And also fixed the testcase exiting failure of when putscom is trying to read from
a register which is having write only access. In this case putscom will return
non-zero return code. So we need to just skip and continue with remaining other
error injections.

this commit fixes the above problem too.

Signed-off-by: Pridhiviraj Paidipeddi <ppaidipe@linux.vnet.ibm.com>