Add test case for pci subsystem info.
This testcase is a start for adding pci or IO subsystem related testcases.
In this version modified the testcase to do comparison of pci devices
between different test results ran against different firmwares with the
known good data of "lspci -mm -n" output of the system when it have
the stable firmware.

Signed-off-by: Pridhiviraj Paidipeddi <ppaidipe@linux.vnet.ibm.com>