Effective execution of all the test cases.
Some test cases will make the system to not to boot (https://github.com/open-power/op-build/issues/672)
Due to that normal power off/on doesn't bring the system, we were end up in other simpler testcases to fail.

This patch makes existing function sys_bmc_power_on_validate_host to below changes.
if system ping failed---> Do IPMI Power off--->Do BMC Cold reset----> Do IPMI Power ON---> wait for system to boot.

And this sequence will executes only when the system is down and if it is up this function will be just
bypassed.

And this function added in all the test cases begining.

So now we have a effective execution of *most* of the test cases at the cost of *time*

Any new testcase addition just use this function at the begining to execute that test case most of the times.

Still if that function doesn't bring the system up, some thing goes wrong in BMC/Hostboot/HW level  ***bug***.

Signed-off-by: Pridhiviraj Paidipeddi <ppaidipe@linux.vnet.ibm.com>