op-ci-basic: start moving to Python unittest based
This avoids the op-ci-bmc-run wrapper (a perl script that constructs
a python script to run the python test that calls the function that actually
does the test).

It also means you can easily execute an individual test case as a standalone
operation.

e.g. to execute the boot and wait for BMC to say we're in working state test,
if you have a valid configuration file, you can just execute:
 ./ci/source/op_ci_bmc.py TestIPL.test_ipl_wait_for_working_state

We currently retain the existing methods and scripts as not all tests
have been converted over.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>