Fix run command on ipmi console.
Temporarily add sleep time of 2 sec's to make sure all other tests use this
function as expected. because so many tests are using this function

If we remove the sleep it is giving some in-correct o/p.

TODO: Remove the sleep and make sure this function returns correct o/p
      for all the command executions.

Signed-off-by: Pridhiviraj Paidipeddi <ppaidipe@linux.vnet.ibm.com>