Remove redundant code and make fwts simpler.
So one can execute fwts tests both remotely through op-test-framework
and in-band also.
Through in-band op-test-framework//testcases/fwts_execution.sh
one can clone and execute the above script.
And changed python code to shell script for easy handling.

Currently it works in Ubuntu Host OS, Not sure this will support
in RHEL OS, if supports that is a TODO.

Signed-off-by: Pridhiviraj Paidipeddi <ppaidipe@linux.vnet.ibm.com>