Major refactor of how tests are run
Initial porting over is only of OpTestPCI and OpTestFWTS,
others will follow in future commits.

Ideas behind this refactor:
- Simplify how to run an individual test or group of tests
- Track state of machine we're testing, to minimise reboots
- be a bit more standard in our use of Python

Now, the top level op-test script is what runs tests. It's a simple
python script that sets up a configuration and what tests to run,
and then just runs them.

Tests are organised into 'suites'. These are a collection of TestSuites
(which are themselves a collection of TestCases and TestSuites).

We use 'suites' in op-test to differentiate between things that an end
user may want to test in a batch. e.g. 'regression' suite,
a suite of tests that are all run at the petitboot prompt, a suite that
needs the OS to be running, etc.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>