OpTestHost: use pxssh instead of manually spawning ssh
We keep a persistent SSH connection, which *dramatically* decreases
test execution time.

For example testcases.OpTestEM.OpTestEMcpu_req_states goes from taking
a total of 44 seconds, down to only 22 seconds.

We also get to remove a couple of weird hacks we had in OpTestFWTS
and OpTestPCI in order to get the real result of a command back.
We do this by doing a bit of echo removal for interacting with ssh

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>