Refactor out tests to use common console interface for IPMI
this (along with the last few patches) get us a pretty standard
way to run a command via either IPMI or an SSH connection.

This allows us to now, in subsequent patches, make code paths
common for running tests over IPMI or SSH (e.g. to easily run the
same test code in petitboot or an OS) *or* to have a different method
of connecting to the out of band host console (e.g. the SSH interface
provided by OpenBMC)

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>