Enable FSP BMC platform enough to run OpTestFastReboot
If --bmc-type FSP is selected, we create a OpTestFSPSystem() which
behaves differently enough to be able to run some level of tests
against an FSP based system.

This enables us to run the FastReboot test as a starting point.

More fixes are needed to run the entire suite, but this is a starting point.

we have to work around looking for a specific skiboot log message, as on
FSP systems we don't get these to the IPMI console. An improvement to the
test is to grep for the message in the OPAL log and ensure the number
of times we've done a fast reboot has increased.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>