Initial OpenBMC support
This gets us to the point of being able to run the skiroot test
suite against an OpenBMC controlled machine that's sitting at
the petitboot shell.

Notably, this patch moves the host console login and unique prompt
setting out from OpTestIPMI up into OpTestSystem, to act against
the console we get from the system's BMC object.

Arguably what should instead be done is that there be a HostConsole
base class with this functionality in it.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>