run_command always return array of output, throw error on errno!=0
This does a final rework to make in band and out of band run_command
look the same to callers. Previously, you would get either a string
of all lines or an array with each element being a line.

We settle on an array of lines as this means that for things like IPMI,
where you may (unexpectedly) get \r\n rather than \n, everything ends up
looking the same.

We also *always* now run "echo $?" after a command to check the exit
code of the command we just ran. In the case where the exit code is not 0,
we raise the (new) CommandFailed() exception with the appropriate information.

The Testcases are appropriately modified to deal with this, including test
cases that previously did *not* check the result of the commands run.
This also allows us to simplify several tests to just iterate over an array
of commands to try to run.

Additionally, in order to make this kind of functionality a bit more obvious,
a few small tests are added and a few misc cleanups throughout the existing
test cases.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>