Use sudo to get host root shell (if needed)
We need to do this in two places: SSH connection and OOB console.

Why? Some distros (notably Ubuntu) really don't want to make it easy to
enable the root user and have them log in via SSH. This is a very good
idea, root has all the power after all. However, we want a root shell
to run tests, so we need to get there somehow. Using sudo is probably
the best current way. We launch a shell rather than using sudo on each
command purely for lazy reasons.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>