testcases/OpTestEEH: Fix some grep & lspci failures.
There can be no error logs in OPAL that means grep will return
non-zero return code, so test will fail. And also when we inject
errors on PCI devices immediatly we are executing lspci that makes
some times failed to read PCI config space of the device where the
error was injected. then lspci also will return non-zero return code.

So for this run_command_ignore_fail implemented which we can use in
those situations where we just want to execute or get the output and
don't care the return code.

So this API fixes those errors in OpTestEEH testcase.

Signed-off-by: Pridhiviraj Paidipeddi <ppaidipe@linux.vnet.ibm.com>