Fix system standby state issue in state machine.
In boston/supermicro systems don't have the Host Status
sensor to check whether system is in standby/runtime state.
So state machine brokes for those systems when system goes
from one state to another state.

This patch fixes this issue by skipping the check when the
given sensor is not available. assumption is system goes to
OFF state.

Signed-off-by: Pridhiviraj Paidipeddi <ppaidipe@linux.vnet.ibm.com>