OpTestIPMI: fix close console path.
There may be situations State machine having the IPMIConsole
is in connected state but SOL console may drop in between.
So close console will try to close the non-existent console session.

So this patch fixes this issue by checking for the active
ipmi sol console session. this function closes the console
only when the session is active.

Signed-off-by: Pridhiviraj Paidipeddi <ppaidipe@linux.vnet.ibm.com>