OpTestPrdDaemon: Fix the command order.
For testing prd daemon, we are killing first and then checking
back whether it is re-spawned back or not. But currently
instead of checking we are againg killing the previous process
which was already killed.

This patch fixes that issue by properly checking the activation
or re-spawn of the process.

Signed-off-by: Pridhiviraj Paidipeddi <ppaidipe@linux.vnet.ibm.com>