test: Add OpTestPNOR to test host pnor access
The host is provided access to its pnor image through an mtd interface
which we can access in petitboot through the pflash utility.

Add a test file OpTestPNOR which contains various tests to check the
ability of the host to modify and update the pnor contents.

The implemented tests are as follows:

- Read, erase and write a partition (NVRAM)
- Overwrite a partition with the existing contents (PAYLOAD/Skiboot)
- Write to the partition table information (toc/part)

Signed-off-by: Suraj Jitindar Singh <sjitindarsingh@gmail.com>