Filter out control characters when logging to console
We will get a log file in test-results/ of all console activity
in the raw, but will now use 'cat -v' to escape control characters
that we print to stdout that we get from the host. This means that
the petitboot UI won't draw over your console when running 'op-test'.

We have a (in progress) '--quiet' option too.

Remove remnants of FFDC capture that didn't capture anything useful.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>