Fix failure of getting the core id's for GARDed cores.
If any Core in the system is garded, host_get_cores fails to get
list of core id's as pir for GARDed core will not be populated in
sysfs. Fix this by skipping those GARDed cores(as error injection
will not happen on those cores).

With this HMI Core checkstop(malfunction test) tests can be run
to test more GARDed cores or to make cores GARDed.

Signed-off-by: Pridhiviraj Paidipeddi <ppaidipe@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>