InstallUbuntu: New test to install Ubuntu
Currently works in Qemu and on host.

Paths to netboot installation are currently hardcoded.

We server up the preseed file and (if host), the netinstall
kernel/initrd from a tiny web server we run as part of the
test.

Additionally, we add some infrastructure for specifying an
Ubuntu install CD and network proxy for accessing the internet
(if doing a net install, that's going to be needed)

We're not yet *booting* that installed OS though,
need to work around some fun with disk labels and partitioning :/

Example of running it:

./op-test ... --run testcases.InstallUbuntu.InstallUbuntu \
  --proxy http://proxy:3128/ \
  --host-scratch-disk /dev/disk/by-id/ata-ST1000NM0033_00LY398_00LY393IBM_Z1W4ZSTB

If --proxy is specified, this will be the proxy used in the installer
and OS to access repositories.

The --host-scratch-disk parameter is the path to a device
(that is recognisible in petitboot *AND* the installer, so by-id is
best) that WILL BE ERASED COMPLETELY. This is whele we'll
install distros onto.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>