Fix BMC FW image flash test according to OpenBMC Implementation.
OpenBMC just does a no operation when updating a code level of 'X'
on a system which is having already code level 'X'.

So we need to follow below procedure to have all scenarios.

1. User requested for flashing image ID 'X'
2. Upload the BMC image.
3. Get the list of BMC image ID's.
4. If 'X' is found in the list
   ---> check if it is Active or Ready
	if it is Ready ---> Activate the image and do a BMC reboot
	if it is already Active
	---> Check whether it is functional(BMC boot side is 'X') or not.
	     if it is functional ---> return
	     if it is non functional ---> Set it's priority to 0, do a BMC reboot
					  and make it functional
5. At the end verify whether BMC boots in right code which we requested for flash.

The idea is to do a no op when updating code 'X' on having system which also has
level of 'X'

Signed-off-by: Pridhiviraj Paidipeddi <ppaidipe@linux.vnet.ibm.com>
[stewart: don't wait quite as long in loop waiting for images to be available]
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>