Deal with IPL/boot failures when booting OS
For example, some bugs may trigger a checkstop during boot of an OS
and not just of petitboot (e.g. when loading NVidia driver when
NVLink hasn't been set up by firmware correctly).

We now catch these properly, change the state of the machine to
UNKNOWN when we do, and catch errors in the FWTS test case.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>