OpTestSystem: Detect machine state
Auto detect the current state of the machine.

When performing tests the current state of the machine
will be checked to transition to the desired state
requested by the test fixture.

This relates to the --machine-state command line option,
instead of having to manually identify the --machine-state
the detection will query the current machine to
see if it can be determined if at Host OS or Petitboot.

If not able to determine the current state then proceed
as usual by entering UNKNOWN machine state.

Signed-off-by: Deb McLemore <debmc@linux.vnet.ibm.com>