TrustedBoot: Fix a broken testsuite failure because of enabled TPM policy
Currently we are enabling the TPM policy at the end(in tearDown) which
leads to a boot terminate when system not having TPM chip and secure mode
enabled. So we need to disable TPM policy only in this case otherwise HB
expects a functional TPM which leads to boot failure in remaining tests

And also documented TPM policy definition which got it from Nick Bofferding
so it will be useful for all who tests trustedboot.

Signed-off-by: Pridhiviraj Paidipeddi <ppaidipe@linux.vnet.ibm.com>