OpTestEM : Monitor time instead of usage to test cpuidle driver
After entering into idle state, a cpu will wakeup in case of a interrupt.
Usage gives number of times a particular state was entered. Thus a cpu
which has more usage may not necessarily have spent more time in idle
state. So time should be used for monitoring of entering and exiting out
of idle state.

Signed-off-by: Abhishek Goel <huntbag@linux.vnet.ibm.com>