op-test: Run InstallHost during DefaultSuite if scratch disk available
If the default test suite is run and a scratch disk and installation
source are available, install the provided OS once the Skiroot suite is
complete. This gives us a known environment once the Host suite starts.

This somewhat naively determines which OS is to be installed based on
the repo or ISO name. If it can't determine the name, installation is
skipped.

Fixes #203

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>