InstallUtil: Fix failure of configure N/W when ping_network fails
get_server_ip tries to ping first, if that fails, then it calls
assign_ip_petitboot and then re-check for ping_network.

But ping_network fails to raise exception when the re-try count
reached to one. Due to which n/w configuration is not triggered
properly.

This fixes by raising a CommandFailed exception when re-try count
reaches to one.

Signed-off-by: Pridhiviraj Paidipeddi <ppaidipe@linux.vnet.ibm.com>