PCI link problem detection
Split the optimalLink function to optimalSpeed and optimalWidth for
extra information to display in assert failure message.

The new subLinkInfo function returns a string containing details of a
suboptimal link.

suboptimal_links is no longer a list of (device, endpoint) tuples. It
now a string concatenation that uses the new subLinkInfo function.

Assert statement still checks that suboptimal_links len() is 0 but now
prints suboptimal_links if it fails.

Signed-off-by: Brenda Lee <lbrenda@au1.ibm.com>