OpTestQemu: Fix disk/cdrom attach
Recent updates to Cedric's powernv Qemu model (which is what we're
realistically running) change how devices must be attached.
Three PHBs are initialised by default but no default PCI layout is
provided[0]. This leaves us with one free slot on each PHB to use.
This can be expanded by adding bridges but for our current uses we
only need two slots, one for the scratch disk and one for the CDROM.

[0]: https://github.com/legoater/qemu/wiki/PowerNV#branch-powernv-30

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>