OpTestQemu: Support attaching full PNOR
QEMU powernv can support attaching a full PNOR image rather than the
individual skiboot and skiroot images. If --host-pnor is specified use
this as the PNOR image to provide to QEMU.
This also requires some fixups to make sure we set the start state when
using QEMU to UNKNOWN_BAD, otherwise OpTestSystem tries to set up the
console too early.

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>