OpTestQemu: Fail test if Qemu command fails
Currently if the Qemu command fails early op-test will loop for a while
trying to "reconnect" by running the command again. If Qemu has failed
this early however it's most likely due to an error in the qemu options
or some internal Qemu error and there's no point waiting for it to fix
itself.
Check the status of the command in connect() and raise a CommandFailed
exception if it has failed immediately. Also remove the reconnect loop
in get_console() as it is not useful.

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>