Pull gcov data at end of skiroot and host test runs
This is an interim solution until we can do this magically by obvserving
a state transition request and going "oh, we need to pull gcov data".

But, this commit will pull into the test run log directory a dump of the
gcov data from skiboot for both the skiroot and host suites. With this
we can get a code coverage report of what code is being covered by
op-test.

Signed-off-by: Stewart Smith <stewart@linux.ibm.com>