OpTestModel: Add OpCheck class and OpTestModel testcase
Add a new helper class OpCheck.  This helper class will allow
testcases which instrument the testcase with the OpCheck class
helper to dynamically perform system wide health checks when
the system wide switch is turned on.

Adding --check-up as a switch to the op-test invocation will
enable the instrumentation to perform some pre and post
testcase class checks dynamically.

When --check-up is added to the op-test arguments, additional
output files will be placed in the logdir for analysis.  When the
--check-up switch is enabled the op-test exit code may be
altered if health checks identify issues to be further investigated.
OpExit code of 117 (errno.EUCLEAN).  <logdir>/opcheck_list* will
provide a summary of the data collected and suggestions for review.

Currently the esel, dmesg and msglog are the common checks performed
(if the system is in a state to collect, e.g. if system is OFF,
no dmesg or msglog are able to be collected).

When --check-up is not explicitly added, no additional checks are
performed nor will the op-test exit code be modified.

OpTestModel testcase is added as an additional resource and
implementation of the OpCheck class helper, as well as illustrating
how to perform custom exploitation of the methods available.

Signed-off-by: Deb McLemore <debmc@linux.ibm.com>