OpTestSystem: Improve Petitboot transitions
This appears to be more resilient when trying to determine where in
Petitboot the machine is, and improves the UI to shell transition.
The changes are;
- Use Ctrl+C in detect target instead of sendline(), which could enter
submenus if in the Petitboot UI.
- Include a call to petitboot_exit_to_shell() in run_PETITBOOT() for the
PETITBOOT state; this causes an explicit drop to shell and return to menu to
verify our position instead of getting lucky with just
exit_petitboot_shell() where the 'x' in 'exit' would happen to exit the
menu.
- Use get_petitboot_prompt() in run_PETITBOOT_SHELL() for the
PETITBOOT_SHELL state; we should already be in the shell.
- Use Ctrl+C in petitboot_exit_to_shell() to have a reliable exit from
the menu. Previously this also worked by chance.
- Explicitly check the current state in setup_term() and return if in
the Petitboot menu.

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>