README: Remove old xmlrunner
Found a bug in the pip install xmlrunner where failures were being
reported as errors.

While investigating where to submit the fix it seems that
unittest-xml-reporting is the proper python package
which properly reflects failures and errors.

To document the problem found if installed the pip xmlrunner:

line 184 /usr/lib/python2.7/site-packages/xmlrunner/xmlrunner.py
        testinfo = _TestInfo(self, test, _TestInfo.ERROR, err)
        self.errors.append((

should be:
        testinfo = _TestInfo(self, test, _TestInfo.FAILURE, err)
        self.failures.append((

So we update op-test README to only suggest unittest-xml-reporting

Signed-off-by: Deb McLemore <debmc@linux.ibm.com>