OpTestQemu: Cleanup temp disk file
When exit path is via optest_handler (via a signal) we need to
cleanup the temp disk file created and/or opened.

Add the qemu_scratch_disk to the cleanup routine.

Add a new qemu_scratch_disk option (overloading the
host_scratch_disk does not allow for proper handling
since the Install* tests require other parameters) and
we need the temp or consumer provided qemu_scratch_disk
to properly close.

Signed-off-by: Deb McLemore <debmc@linux.ibm.com>