OpTestSystem: Aide PTY tracing
Add some debug tracing to aide during problem determination
for pexpect pty buffer management.

When the pty object gets closed, dies, or some other state that
renders the pty object not usable, then the old pty buffer and sockets
get closed and any data waiting or pending on the old socket is lost.

A new pty object is created (new pexpect spawn of pty process, etc)
and this new process and new buffer are empty upon creation.

Until new data arrives on this new pty object buffer there is nothing
to be read.  The application writer (you!) will need to either kick
something in the new buffer or know that some future action will
pipe data to the new socket so that things are not hung (timing issues
can make this unpredictable), so its best that an explicit action
occurs, e.g. sendline "date".  This methodology requires that
you are at a predictable location and predictable reactions
can occur.

Signed-off-by: Deb McLemore <debmc@linux.ibm.com>