OpTestUtil: Check if already root before trying sudo
Make a call to check_root() to see if the console is already in a root
shell. This avoids nested shells where we don't expect it, like in the
Petitboot shell which in newer versions has sudo available, and means we
don't set up the prompt two times in a row.

Also remove two variables which are set but unused.

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>