Add FastRebootHostStressTorture: test fast-reboot while host is stressed
There's been some reports of some issues with SRESET on POWER8 while the
host is under load.

This is an attempt to test for any such problem by using the 'stress'
utility to produce some workload on the host and then do fast reboot.

This currently (as of skiboot v6.2-176-g261ca8e779e5) fails on POWER8
pretty spectacularly, which seems to be the same thing that was reported by
YADRO in https://github.com/open-power/skiboot/issues/185

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>