Put error messages on one line
Error messages are helpful. They are doubly helpful when a user can grep
the source tree for the error string and find where they are. Splitting
the message over multiple lines breaks this.

If someone has a PEP8 compliant variant of this commit I'd be happy to
follow their example, as long as the strings are unbroken.

Signed-off-by: Joel Stanley <joel@jms.id.au>