common: Improve Petitboot detection and shell exit
Commit 97b808e3 "OpTestSystem: Improve Petitboot transitions" made
several improvements to reliably dealing with the Petitboot menu. Part
of this was using Ctrl-C to exit the UI. While this is better for making
sure the UI is exited, it drops the system to a less functional shell
than the one normally started by exiting Petitboot normally.

Return to using 'x' to exit the UI but properly this time and with a bit
more paranoia. Instead of using sendline() or Ctrl-C to drive output,
use Ctrl-L and handle the extra 'erase' character that frustrated
previous efforts at its use. In petitboot_exit_to_shell() an 'x' is sent
on each attempt to gain the prompt in case the system is currently in a
Petitboot submenu.

This maintains reliability when dealing with the Petitboot state and
avoids issues with Ctrl-C and similar signals when dropped to a shell.

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>