OpTestUtil: Trigger refresh of host login prompt
When logging into the host OS out-of-band (eg. IPMI) the login prompt
does not respond to Ctrl-L as expected. If we know we're waiting for a
host login prompt (eg. we know the state or we're booting up from
Petitboot), trigger a proper refresh by hitting enter at the prompt.

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>