OpTestSystem: Force prompt refresh if unresponsive in detect_target
Similar to b8629a8 "OpTestUtil: Trigger refresh of host login prompt"
the same problem is present in detect_target(). detect_target() is
called in a loop from run_DETECT() - if it fails to detect a state the
first time, send a carriage return to force the host login prompt to
print since it doesn't respond to Ctrl-L.

Additionally if detect_target() fails to detect a state it increments
detect_counter, however the loop in run_DETECT() also does this and the
detect loop only ever executes once. Remove the increment in
detect_target() and leave it to run_DETECT().

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>