util: Make hostlocker config parsing more robust
The parsing of op-test configuration from hostlocker comments is prone
to breaking since we just scan for a config key + ":" in the line. I've
noticed that op-test will treat a line that ends with "Run:" as an op-test
configuration directive which caused the test to fail since the rest
of the line didn't define a valid test case.

This is kind of stupid, so make the parsing more robust by:

a) Making op-test scan for an "[op-test]" marker that indicates when the
   op-test config definitions begin.

b) Only treating a line as an op-test config option if it begins with
   "key:" rather than allowing it to be present anywhere in the line.
   This has the side effect of allowing you to comment out a directive.

Additionally, this makes the configuration logged into the debug log,
with the exception of passwords which are logged as "<hidden>"

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>