Refactor the entire codebase to be pep8 compliant
PEP 8 is the standard Python style.  There is support for it in
virtually every Python linter and text editor, which will format your
code for you, or point out when something is noncompliant.

This is a *gigantic* diffstat, of which I personally wouldn't consider
every change (not that I've manually reviewed them all) an improvement.
However, it fixes a huge amount of inconsistent whitespace, misleading
indents, duplicated imports and definitions, etc.

It's trivial to set up PEP8 checkers that run on every pull request, so
that we can keep the code style in op-test consistent.  Let's bite the
bullet here, and keep that standard maintained going forward.

Signed-off-by: Russell Currey <ruscur@russell.cc>