Convert the codebase to Python 3
Python 2 will be deprecated on January 1, 2020.  Most Linux
distributions have either already switched to Python 3 as default, or
are planning to soon.

Most changes here are just renamings or basic syntax, thankfully most
op-test code is not particularly problematic here, outside of pexpect.

The only manual changes made other than running 2to3:

  - removed the "async" arg from expect()
      it was unused and async is a keyword in Python 3
  - no need to set the default encoding to UTF-8 in ./op-test
  - ignore "__pycache__" directory in addons/
  - convert some strings to byte types
  - rename __del__() to cleanup()
      python3 prevents you from calling open() in __del__ handlers,
      which happens if you try to log.  we already register these with
      atexit, which works fine
  - call pexpect.spawn with encoding='utf-8'
  - open logfile with text=True
  - petitboot-i18n converted to *actually* unicode

There WILL be breakages, please report them or fix them.  This is a
necessary change and the sooner we do it, the better.

Signed-off-by: Russell Currey <ruscur@russell.cc>