OpTestQemu: Add PCI bridges to support more devices.
The powernv QEMU model only has 3 free slots which doesn't leave a lot
of room for tests that want more complicated setups.
Instead fill each of these slots with a PHB which we can insert further
devices into.

This also extends the scratch disk concept to support an arbitrary
number of disks. The Petitboot10000Disks testcase provides an example of
this functionality. The intended use case for this is to extend the
Petitboot testcases to test against different source devices,
filesystems, and configuration formats.

For particularly ambitious testcases the QEMU setup will also
dynamically insert more PHBs into the model to support an arbitrary
number of devices.

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>