PetitbootConfig: Update configuration tests
A bunch of configuration-related tests were waiting for NVRAM support
and a reliable network address in Qemu. Now that we have it, update
these tests so that
a) they run at all,
b) they reflect current Petitboot behaviour, and
c) try to preserve the state of NVRAM before the test.

This has also had autopep8 run against it so bring it in to line with
other work.

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>