OpTestFlash: Rework get_version_tar() to suck less
The previous code would iterate over every file in the tarball until it
found "version=", and then go back over the file it's in, and parse it
out from there.  Instead of doing that and potentially going through the
entire PNOR blob, directly check the MANIFEST.

This should also fix any issues with Python 3 string encoding.

Signed-off-by: Russell Currey <ruscur@russell.cc>