[v2] Refactor PowerVM code aligning to existing framework
This patch refactors existing PowerVM specific code implementation
to be aligned with the other BMC types i.e. irrespective of the
BMC type.

An LPAR is considered a SUT equivalent to a PowerNV machine.

This patch helps us achieve a common interface of accessing and
using SUT (LPAR or PowerNV). Methods which are not applicable for
LPARs are overridden wherever possible.

v2 Changelog:
1) Get console connected initially for test to run commands with
2) Detect initial poweroff and recover system for use

Fixes: #543

Signed-off-by: Harish <harish@linux.ibm.com>