Add support to check if kdump status and check for failure
Added function (host_is_kdump_enabled) to check if kdump is enabled or not.
Also added a check to find if start/stop of kdump was success, exit if it is
failed.

Added option --no-pager to systemctl status command, without this option the
command will wait to timeout because it expects "q" to quit.

Signed-off-by: Hariharan T S <hari@linux.vnet.ibm.com>