config: Seperate log and locker setup from config parsing
Not strictly required, but we're doing a lot in parse_config()
that has nothing to do with parsing config file. Move it into
OpTestConfiguration.conf.do_testing_setup() which handles
setting up the log files and getting host lockers. We should
probably split that further, but that's a problem for another day.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>